requirements.txt
# Core web/runtime stack
fastapi>=0.112
uvicorn[standard]>=0.30
python-dotenv>=1.0

# OpenAI-compatible client layer
openai>=2.0.0
httpx>=0.27
